Remove "deprecated" metadata on lex-permutations
Can we remove this deprecation? As the symbol is not public anymore, I don't think it is necessary, and is causing warnings when we use it on shadow-cljs.

As shadow-cljs will not reload file with warnings, we need to disable all warnings to be able to reload code, but this is less than ideal because it'll also disable "undefined symbols" and "wrong number of arguments" and other useful info.